// KatinoUI/src/app/core/enums/order-internet-doc-status.ts
export enum OrderInternetDocStatus {
  notProcessed = 0,
  created = 1,
  deleted = 2,
  notFound = 3,
  inTheCityInterregional = 4,
  onTheWayToCity = 5,
  onTheWayToDepartment = 6,
  arrived = 7,
  arrivedPostomat = 8,
  received = 9,
  receivedRemittancePending = 10,
  receivedRemittanceCompleted = 11,
  npCompletingOrder = 12,
  inTheCityWithinTheCity = 41,
  onTheWayToReceiver = 101,
  rejectionBySender = 102,
  rejection = 103,
  addressChanged = 104,
  storageStopped = 105,
  receivedAndReturnCreated = 106,
  receiverNotAnswering = 111,
  deliveryDateChangedByReceiver = 112,
}